Initial version of End game mode implemented. It's not great, and needs a refactor. But it works.
svn path=/trunk/bitsharp/; revision=67495

+++ src/MonoTorrent.Client/PiecePicking/IPiecePicker.cs
@@ -13,6 +13,6 @@ namespace MonoTorrent.Client
         int CurrentRequestCount();
         void ReceivedRejectRequest(PeerConnectionID id, RejectRequestMessage message);
         void RemoveRequests(PeerConnectionID id);
-        void ReceivedPieceMessage(PeerConnectionID id, byte[] buffer, int dataOffset, int writeIndex, int blockLength, PieceMessage message);
+        void ReceivedPieceMessage(PeerConnectionID id, byte[] buffer, int dataOffset, long writeIndex, int blockLength, PieceMessage message);
     }
 }
